refactor: move SDK custom fields to separate file
