+++ projects/stream-chat-angular/src/stream-chat.ts
@@ -1,5 +1,5 @@
 import 'stream-chat';
-import { DefaultChannelData, DefaultAttachmentData } from './lib/types';
+import { DefaultChannelData, DefaultAttachmentData } from './lib/types-custom';
 
 declare module 'stream-chat' {
   interface CustomAttachmentData extends DefaultAttachmentData {}
